requirements.txt
ttkbootstrap>=1.10.1
psutil>=5.9.0
pystray>=0.19.4
Pillow>=10.0.0
pywin32>=306
